fix: issue 371
